move strategies to separate files
